feat: 드롭다운 외부 클릭 시 닫히도록 기능 추가
